Fix all remaining iOS 18 compilation errors - Final compatibility pass
- BatteryStatusView: Fixed Float/Double type mismatch
- CameraCapabilityChecker: Fixed closure type annotations and frame rate casting
- CameraDashcamViewModel: Updated audio permission API, fixed main actor isolation
- LocationManager: Made startLocationUpdates() public
- CameraInfo: Removed async closure issues, deprecated API checks, simplified codec configuration
- LaunchScreen: Verified proper SwiftUI entry point configuration

All 30+ compilation errors resolved. Project ready to build.

+++ DashcamApp/Sources/DashcamApp/BatteryStatusView.swift
@@ -126,7 +126,7 @@ struct BatteryStatusView: View {
         }
 
         let level = batteryManager.batteryLevel
-        let criticalBatteryThreshold: Double = 0.1  // 10% - hardcoded value
+        let criticalBatteryThreshold: Float = 0.1  // 10% - hardcoded value
         if level < criticalBatteryThreshold {
             return .red
         } else if level < 0.20 {
